backend/requirements.txt
fastapi
pydantic
pydantic-settings
uvicorn
google-genai